requirements.txt
# Core Dependencies
torch>=1.9.0
transformers>=4.20.0
numpy>=1.21.0
pandas>=1.3.0

# Data Processing
scikit-learn>=1.0.0

# Configuration
PyYAML>=5.4.0

# Visualization (for notebooks)
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.62.0

# Optional: For better performance
 accelerate>=0.20.0
datasets>=2.0.0

